pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tool-inventory"
version = "0.6.0"
description = "Tool inventory."
authors = ["Micael Jarniac <micael@jarniac.dev>"]
license = "MIT"
readme = "docs/README.md"
packages = [{include = "tool_inventory", from = "src"}]
homepage = "https://github.com/MicaelJarniac/tool-inventory"
repository = "https://github.com/MicaelJarniac/tool-inventory"
documentation = "https://tool-inventory.readthedocs.io"
classifiers = [
  # Get more from http://pypi.python.org/pypi?%3Aaction=list_classifiers
  "Development Status :: 1 - Planning",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]

[tool.poetry.scripts]
tool-inventory = "tool_inventory.cli:app"

[tool.poetry.dependencies]
python = "^3.13"

loguru = "^0.7.0"
sqlmodel = "^0.0.22"
fastapi = {extras = ["standard"], version = "^0.115.5"}
thefuzz = "^0.22.1"
typer = "^0.14.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
fastapi-users = "^14.0.0"
fastapi-users-db-sqlmodel = "^0.3.0"

[tool.poetry.group.tests.dependencies]
coverage = "^7.2.7"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-benchmark = "^4.0.0"

[tool.poetry.group.benchmarking.dependencies]
asv = "^0.5.1"
virtualenv = "^20.24.3"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = {extras = ["python"], version = "^0.27.0"}
mkdocs-material = "^9.5.46"
pymdown-extensions = "^10.12"
mkdocs-gen-files = "^0.5.0"
mkdocs-section-index = "^0.3.9"
mkdocs-literate-nav = "^0.6.1"
griffe-pydantic = "^1.1.0"

[tool.poetry.group.typing.dependencies]
# Typing stubs go here
mypy = "^1.4.0"
nox = "^2023.4.22"
pytest = "^7.4.0"
pytest-benchmark = "^4.0.0"

[tool.poetry.group.linting.dependencies]
ruff = "^0.8.0"

[tool.poetry.group.pre-commit.dependencies]
pre-commit = "^3.3.3"

[tool.poetry.group.dev.dependencies]
nox = "^2023.4.22"
python-semantic-release = "^8.0.2"

[tool.poetry.urls]
"Changelog" = "https://github.com/MicaelJarniac/tool-inventory/blob/main/docs/CHANGELOG.md"
"Bug Tracker" = "https://github.com/MicaelJarniac/tool-inventory/issues"

[tool.ruff]
# https://beta.ruff.rs/docs/rules
line-length = 88  # Default
force-exclude = true
src = ["src"]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D203",
  "D213",
]
unfixable = ["ERA001"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id", "type"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
files = ["src", "scripts", "tests", "noxfile.py"]
exclude = [".venv"]
enable_error_code = [
  "ignore-without-code"
]
strict = true
warn_unused_ignores = false  # true
show_column_numbers = true
# show_error_context = true
# pretty = true

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
build_command = """
  python -m pip install build~=0.10.0
  python -m build .
"""

[tool.semantic_release.changelog]
changelog_file = "docs/CHANGELOG.md"

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.rc]
match = "rc"
prerelease = true
prerelease_token = "rc"

[tool.coverage.run]
source = ["src"]
command_line = "-m pytest"

[tool.pytest.ini_options]
# addopts = "--cov"

[tool.cruft]
skip = [
  "src",
  "tests",
  "benchmarks",
  ".asv",
  "poetry.lock",
]
